move `AccentColorListener` to separate file
